orbit-hub/backend/requirements.txt
fastapi
uvicorn[standard]
google-genai

